As you can see
there are lots of changes and additions. i added switch part which make special cards controlable when the user wants use them. and i added the hold and restart part. restart part is not complicated. its hold set points and after any set win it restart the all values and create the object again.
In the hold part there will be changes so im not going to talk about it.

Last, in the playerDeck class, i change the some values names. Not important